replace tables + iterators with sequins
during a chat with jonathan snyder about the upcoming workshop, he brought up that it'd be rad to expose learners to sequins early on, since it allows beginners to achieve musical results without much prerequisite knowledge about Lua tables. script runs the same as the previous version. this update coincides with an update to the actual study text.